+++ validator-core/src/main/java/fr/ign/validator/validation/attribute/GeometryDataExtentValidator.java
@@ -19,7 +19,7 @@ public class GeometryDataExtentValidator implements Validator<Attribute<Geometry
 
 	@Override
 	public void validate(Context context, Attribute<Geometry> attribute) {
-		Geometry geometry = attribute.getValue() ;
+		Geometry geometry = attribute.getBindedValue() ;
 		
 		if ( null == geometry || geometry.isEmpty() ){
 			return ;