+++ validator-core/src/test/java/fr/ign/validator/model/type/UrlTypeTest.java
@@ -1,10 +1,8 @@
 package fr.ign.validator.model.type;
 
-import org.junit.Test;
+import java.net.URL;
 
-import fr.ign.validator.Context;
-import fr.ign.validator.report.InMemoryReportBuilder;
-import junit.framework.TestCase;
+import org.junit.Test;
 
 
 /**
@@ -12,29 +10,16 @@ import junit.framework.TestCase;
  * @author CBouche
  *
  */
-public class UrlTypeTest extends TestCase {
-	private UrlType attribute;
-	protected Context context ;
-	protected InMemoryReportBuilder reportBuilder ;
-	
+public class UrlTypeTest extends AbstractTypeTest<URL> {
+
 	public UrlTypeTest() {
-		attribute = new UrlType() ;
-		attribute.setName("test");
-		
-		context = new Context() ;
-		reportBuilder = new InMemoryReportBuilder() ; 
-		context.setReportBuilder(reportBuilder);
+		super(new UrlType());
 	}
 	
 	@Test
 	public void testCheckAttributeValue() {
-//		Properties systemSettings = System.getProperties();
-//		systemSettings.put("proxySet", "true");
-//		systemSettings.put("http.proxyHost", "proxy.ign.fr");
-//		systemSettings.put("http.proxyPort", "3128");
-		attribute.bindValidate(context,"ce n'est pas une url") ;
+		bindValidate(context,"ce n'est pas une url") ;
 		assertFalse( reportBuilder.isValid() );
-//		assertTrue( property.validate("http://www.google.fr/") );
 	}
 	
 }