+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * [x] add option ```--plugins``` and remove maven profile "cnig"
 * [x] simplify error context management
 * [x] check ogr2ogr's version and ban some versions
+* [x] escape control characters in error report
+* [x] validate input characters, report error for non standard control characters and non latin1 convertible characters
+
 
 # 3.1.* (next)